create `page-layout` folder and move some components into it
