+++ products/statement-generator/src/involvement-step/InvolvementUnemployedFlow.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { IInvolvementUnemployedState } from 'contexts/FormStateProps';
 
-import FlowNavigation from 'components/FlowNavigation';
+import FlowNavigation from 'page-layout/FlowNavigation';
 import Textarea from 'components/Textarea';
 import RadioGroup from 'components/RadioGroup';
