+++ products/statement-generator/src/page-layout/AppFooter.tsx
